+++ frontend/src/components/layout/TabBar/TabBar.tsx
@@ -2,9 +2,9 @@ import { useNavigate } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
 
 import gongseek from '@/assets/gongseek.png';
-import UserProfileIcon from '@/components/@common/UserProfileIcon/UserProfileIcon';
-import { UserProfile } from '@/components/@common/UserProfileIcon/UserProfileIcon.styles';
 import * as S from '@/components/layout/TabBar/TabBar.styles';
+import UserProfileIcon from '@/components/user/UserProfileIcon/UserProfileIcon';
+import { UserProfile } from '@/components/user/UserProfileIcon/UserProfileIcon.styles';
 import { URL } from '@/constants/url';
 import { menuSliderState } from '@/store/menuSliderState';
 import { getUserIsLogin } from '@/store/userState';
