+++ frontend/src/components/user/UserCommentBox/UserCommentBox.tsx
@@ -1,8 +1,8 @@
 import EmptyMessage from '@/components/@common/EmptyMessage/EmptyMessage';
 import Loading from '@/components/@common/Loading/Loading';
+import * as S from '@/components/user/UserCommentBox/UserCommentBox.styles';
+import UserCommentItem from '@/components/user/UserCommentItem/UserCommentItem';
 import useGetUserComments from '@/hooks/user/useGetUserComments';
-import * as S from '@/pages/MyPage/UserCommentBox/UserCommentBox.styles';
-import UserCommentItem from '@/pages/MyPage/UserCommentItem/UserCommentItem';
 
 const UserCommentBox = () => {
 	const {
