+++ frontend/src/components/user/UserCommentItem/UserCommentItem.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 
-import * as S from '@/pages/MyPage/UserCommentItem/UserCommentItem.styles';
+import * as S from '@/components/user/UserCommentItem/UserCommentItem.styles';
 import { UserComment } from '@/types/commentResponse';
 import { dateTimeConverter } from '@/utils/converter';
