+++ frontend/src/components/user/UserProfileIcon/UserProfileIcon.styles.tsx
