+++ frontend/src/components/vote/AddedOption/AddedOption.tsx
